Fix persona dialog review feedback
