+++ desktop/src/features/agents/ui/PersonaDialog.tsx
@@ -1,8 +1,6 @@
 import * as React from "react";
 import { RefreshCw, Upload } from "lucide-react";
 
-import { ProfileAvatar } from "@/features/profile/ui/ProfileAvatar";
-import { useAvatarUpload } from "@/features/profile/useAvatarUpload";
 import type {
   AcpRuntimeCatalogEntry,
   CreatePersonaInput,
@@ -14,8 +12,8 @@ import { Button } from "@/shared/ui/button";
 import { ChooserDialogContent } from "@/shared/ui/chooser-dialog-content";
 import { Dialog } from "@/shared/ui/dialog";
 import { Input } from "@/shared/ui/input";
-import { Spinner } from "@/shared/ui/spinner";
 import { Textarea } from "@/shared/ui/textarea";
+import { AgentCreationPreview } from "./AgentCreationPreview";
 import { EnvVarsEditor, type EnvVarsValue } from "./EnvVarsEditor";
 import {
   getImportButtonLabel,
@@ -55,6 +53,7 @@ const PERSONA_LABEL_OPTIONAL_CLASS =
 const AUTO_MODEL_DROPDOWN_VALUE = "__auto_model__";
 const CUSTOM_MODEL_DROPDOWN_VALUE = "__custom_model__";
 const AUTO_PROVIDER_DROPDOWN_VALUE = "__auto_provider__";
+const CUSTOM_PROVIDER_DROPDOWN_VALUE = "__custom_provider__";
 
 type PersonaModelOption = {
   id: string;
@@ -184,6 +183,8 @@ export function PersonaDialog({
   const [model, setModel] = React.useState("");
   const [isCustomModelEditing, setIsCustomModelEditing] = React.useState(false);
   const [provider, setProvider] = React.useState("");
+  const [isCustomProviderEditing, setIsCustomProviderEditing] =
+    React.useState(false);
   const [envVars, setEnvVars] = React.useState<EnvVarsValue>({});
   const [isAvatarUploadPending, setIsAvatarUploadPending] =
     React.useState(false);
@@ -211,6 +212,7 @@ export function PersonaDialog({
     setModel(initialValues.model ?? "");
     setIsCustomModelEditing(false);
     setProvider(initialValues.provider ?? "");
+    setIsCustomProviderEditing(false);
     setEnvVars("envVars" in initialValues ? (initialValues.envVars ?? {}) : {});
     setIsAvatarUploadPending(false);
     setImportErrorMessage(null);
@@ -334,6 +336,7 @@ export function PersonaDialog({
       setModel("");
       setIsCustomModelEditing(false);
       setProvider("");
+      setIsCustomProviderEditing(false);
       setEnvVars({});
       setIsAvatarUploadPending(false);
       setImportErrorMessage(null);
@@ -354,6 +357,11 @@ export function PersonaDialog({
     }
 
     const trimmedRuntime = runtime.trim();
+    const previousRuntime = initialValues.runtime?.trim() ?? "";
+    const shouldPreserveHiddenModelProvider =
+      "id" in initialValues &&
+      previousRuntime.length === 0 &&
+      trimmedRuntime.length === 0;
     const preservedNamePool =
       "namePool" in initialValues ? initialValues.namePool : undefined;
     const baseInput = {
@@ -361,8 +369,16 @@ export function PersonaDialog({
       avatarUrl: avatarUrl.trim() || undefined,
       systemPrompt: systemPrompt.trim(),
       runtime: trimmedRuntime || undefined,
-      model: trimmedRuntime ? model.trim() || undefined : undefined,
-      provider: trimmedRuntime ? provider.trim() || undefined : undefined,
+      model: trimmedRuntime
+        ? model.trim() || undefined
+        : shouldPreserveHiddenModelProvider
+          ? initialValues.model
+          : undefined,
+      provider: trimmedRuntime
+        ? provider.trim() || undefined
+        : shouldPreserveHiddenModelProvider
+          ? initialValues.provider
+          : undefined,
       namePool: preservedNamePool,
       envVars,
     };
@@ -415,6 +431,10 @@ export function PersonaDialog({
   const showCustomModelInput =
     modelFieldVisible && (isCustomModelEditing || isModelCustom);
   const providerOptions = getPersonaProviderOptions(provider);
+  const providerSelectValue = isCustomProviderEditing
+    ? CUSTOM_PROVIDER_DROPDOWN_VALUE
+    : provider.trim() || AUTO_PROVIDER_DROPDOWN_VALUE;
+  const showCustomProviderInput = modelFieldVisible && isCustomProviderEditing;
   const blankRuntimeOptionLabel =
     isCreateMode && runtimesLoading
       ? "Loading providers..."
@@ -621,6 +641,7 @@ export function PersonaDialog({
                     }
                     if (nextRuntime.trim().length === 0) {
                       setIsCustomModelEditing(false);
+                      setIsCustomProviderEditing(false);
                       setProvider("");
                     }
                   }}
@@ -778,13 +799,20 @@ export function PersonaDialog({
                       id="persona-llm-provider"
                       onChange={(event) => {
                         const nextProvider = event.target.value;
+                        if (nextProvider === CUSTOM_PROVIDER_DROPDOWN_VALUE) {
+                          setIsCustomProviderEditing(true);
+                          setProvider("");
+                          return;
+                        }
+
+                        setIsCustomProviderEditing(false);
                         setProvider(
                           nextProvider === AUTO_PROVIDER_DROPDOWN_VALUE
                             ? ""
                             : nextProvider,
                         );
                       }}
-                      value={provider.trim() || AUTO_PROVIDER_DROPDOWN_VALUE}
+                      value={providerSelectValue}
                     >
                       <option value="" disabled>
                         {selectedProviderLabel}
@@ -797,8 +825,33 @@ export function PersonaDialog({
                           {option.label}
                         </option>
                       ))}
+                      <option value={CUSTOM_PROVIDER_DROPDOWN_VALUE}>
+                        Custom provider...
+                      </option>
                     </select>
                   </div>
+                  {showCustomProviderInput ? (
+                    <div
+                      className={cn(
+                        "mt-2 flex min-h-11 items-center px-3",
+                        PERSONA_FIELD_SHELL_CLASS,
+                      )}
+                    >
+                      <Input
+                        aria-label="Custom provider ID"
+                        autoCorrect="off"
+                        className={cn(
+                          "h-8 px-0 py-0 leading-6",
+                          PERSONA_FIELD_CONTROL_CLASS,
+                        )}
+                        disabled={isPending || !modelFieldVisible}
+                        id="persona-custom-provider"
+                        onChange={(event) => setProvider(event.target.value)}
+                        placeholder="Custom provider ID"
+                        value={provider}
+                      />
+                    </div>
+                  ) : null}
                 </div>
               </div>
             </div>
@@ -816,165 +869,5 @@ export function PersonaDialog({
         </form>
       </ChooserDialogContent>
     </Dialog>
-  );
-}
-
-function isAvatarFileDrag(event: React.DragEvent<HTMLElement>) {
-  return Array.from(event.dataTransfer.types).includes("Files");
-}
-
-function AgentCreationPreview({
-  avatarUrl,
-  disabled = false,
-  label,
-  onUploadPendingChange,
-  onSelectAvatar,
-}: {
-  avatarUrl: string | null;
-  disabled?: boolean;
-  label: string;
-  onUploadPendingChange?: (isPending: boolean) => void;
-  onSelectAvatar: (avatarUrl: string) => void;
-}) {
-  const [isDragOverAvatar, setIsDragOverAvatar] = React.useState(false);
-  const avatarDragDepthRef = React.useRef(0);
-  const {
-    inputRef: avatarUploadInputRef,
-    isUploading,
-    errorMessage: uploadErrorMessage,
-    clearError: clearUploadError,
-    openPicker: openUploadPicker,
-    uploadFile: uploadAvatarFile,
-    handleFileChange: handleAvatarUploadFileChange,
-  } = useAvatarUpload({
-    onUploadSuccess: onSelectAvatar,
-  });
-
-  React.useEffect(() => {
-    onUploadPendingChange?.(isUploading);
-    return () => {
-      onUploadPendingChange?.(false);
-    };
-  }, [isUploading, onUploadPendingChange]);
-
-  const handleAvatarDragEnter = React.useCallback(
-    (event: React.DragEvent<HTMLFieldSetElement>) => {
-      if (disabled || !isAvatarFileDrag(event)) {
-        return;
-      }
-      event.preventDefault();
-      event.stopPropagation();
-      avatarDragDepthRef.current += 1;
-      event.dataTransfer.dropEffect = "copy";
-      setIsDragOverAvatar(true);
-    },
-    [disabled],
-  );
-
-  const handleAvatarDragOver = React.useCallback(
-    (event: React.DragEvent<HTMLFieldSetElement>) => {
-      if (disabled || !isAvatarFileDrag(event)) {
-        return;
-      }
-      event.preventDefault();
-      event.stopPropagation();
-      event.dataTransfer.dropEffect = "copy";
-      setIsDragOverAvatar(true);
-    },
-    [disabled],
-  );
-
-  const handleAvatarDragLeave = React.useCallback(
-    (event: React.DragEvent<HTMLFieldSetElement>) => {
-      if (!isAvatarFileDrag(event)) {
-        return;
-      }
-      event.preventDefault();
-      event.stopPropagation();
-      avatarDragDepthRef.current = Math.max(0, avatarDragDepthRef.current - 1);
-      if (avatarDragDepthRef.current === 0) {
-        setIsDragOverAvatar(false);
-      }
-    },
-    [],
-  );
-
-  const handleAvatarDrop = React.useCallback(
-    (event: React.DragEvent<HTMLFieldSetElement>) => {
-      if (!isAvatarFileDrag(event)) {
-        return;
-      }
-      event.preventDefault();
-      event.stopPropagation();
-      avatarDragDepthRef.current = 0;
-      setIsDragOverAvatar(false);
-
-      const file = event.dataTransfer.files[0];
-      if (!file || disabled || isUploading) {
-        return;
-      }
-
-      clearUploadError();
-      void uploadAvatarFile(file);
-    },
-    [clearUploadError, disabled, isUploading, uploadAvatarFile],
-  );
-
-  return (
-    <div className="mx-auto w-full max-w-[220px] lg:sticky lg:top-0">
-      <fieldset
-        aria-label="Agent avatar preview"
-        className={cn(
-          "group/avatar-preview relative m-0 aspect-[4/5] min-h-[240px] min-w-0 overflow-hidden rounded-xl border border-border/70 bg-muted/50 p-0 shadow-xs transition-[background-color,border-color,box-shadow] duration-150",
-          isDragOverAvatar &&
-            "border-dashed border-primary/70 bg-primary/5 ring-2 ring-primary/15",
-        )}
-        onDragEnter={handleAvatarDragEnter}
-        onDragLeave={handleAvatarDragLeave}
-        onDragOver={handleAvatarDragOver}
-        onDrop={handleAvatarDrop}
-      >
-        <input
-          accept="image/gif,image/jpeg,image/png,image/webp"
-          className="hidden"
-          onChange={handleAvatarUploadFileChange}
-          ref={avatarUploadInputRef}
-          type="file"
-        />
-
-        <div className="absolute inset-0 flex items-center justify-center">
-          <ProfileAvatar
-            avatarUrl={avatarUrl}
-            className="h-36 w-36 text-4xl"
-            label={label}
-          />
-        </div>
-
-        {uploadErrorMessage ? (
-          <p className="absolute inset-x-3 bottom-12 rounded-md bg-background/95 px-2 py-1 text-center text-xs text-destructive shadow-xs">
-            {uploadErrorMessage}
-          </p>
-        ) : null}
-
-        <div className="absolute inset-x-3 bottom-3 flex justify-center">
-          <button
-            className="inline-flex h-8 translate-y-1 items-center justify-center gap-1.5 rounded-full border border-border/70 bg-background/90 px-3 text-xs font-medium text-foreground opacity-0 shadow-xs transition-[background-color,opacity,transform] duration-150 hover:bg-muted focus-visible:translate-y-0 focus-visible:opacity-100 focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-ring group-hover/avatar-preview:translate-y-0 group-hover/avatar-preview:opacity-100 group-focus-within/avatar-preview:translate-y-0 group-focus-within/avatar-preview:opacity-100"
-            disabled={disabled || isUploading}
-            onClick={() => {
-              clearUploadError();
-              openUploadPicker();
-            }}
-            type="button"
-          >
-            {isUploading ? (
-              <Spinner className="h-3.5 w-3.5 border-2" />
-            ) : (
-              <Upload className="h-3.5 w-3.5" />
-            )}
-            {isUploading ? "Uploading..." : "Edit avatar"}
-          </button>
-        </div>
-      </fieldset>
-    </div>
   );
 }
